Refactor to src layout with cli/pdf_generator packages and working console script
- Migrated project to a src/ layout for cleaner packaging
- Split code into two packages: cli/ for entrypoints and pdf_generator/ for core logic
- Added pyproject.toml with setuptools configuration and console script definition
- Ensured __init__.py files across all subpackages
- Updated make_pdfs.py to import via pdf_generator.* absolute imports
- Verified editable install works with `pip install -e .`
- Confirmed console script `make-pdfs` executes successfully and generates JSON/PDF pairs

 Changes to be committed:
	modified:   README.md
	new file:   pyproject.toml
	new file:   src/__init__.py
	new file:   src/api/__init__.py
	renamed:    fetch_clickup_task.py -> src/api/fetch_clickup_task.py
	new file:   src/cli/__init__.py
	renamed:    make_pdfs.py -> src/cli/make_pdfs.py
	new file:   src/oauth/__init__.py
	new file:   src/pdf_generator/__init__.py
	renamed:    clickup_parser.py -> src/pdf_generator/clickup_parser.py
	renamed:    generate_pdf.py -> src/pdf_generator/generate_pdf.py
	renamed:    renderers.py -> src/pdf_generator/renderers.py
	renamed:    styles.py -> src/pdf_generator/styles.py
	renamed:    utils.py -> src/pdf_generator/utils.py
	new file:   src/webapp/__init__.py
