+++ pyproject.toml
@@ -1,0 +1,18 @@
+[build-system]
+requires = ["setuptools>=68", "wheel"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "clickup-pdf-generator"
+version = "0.1.0"
+description = "Generate pretty PDFs from ClickUp tasks"
+requires-python = ">=3.10"
+
+[tool.setuptools]
+package-dir = {"" = "src"}
+
+[tool.setuptools.packages.find]
+where = ["src"]
+
+[project.scripts]
+make-pdfs = "cli.make_pdfs:main"
