+++ src/pdf_generator/styles.py
